Move definition of test settings from a fixture to session startup
This fixes a sneaky class of setting leaks where the test settings are not applied during collection of tests. This means that if you instantiate an object in a set of parameters for a parameterized test, it could use non-test settings. This can allow an object defined in tests to perform unintentional actions on the real PREFECT_HOME directory.