Install the `prefect` package into the environment when testing subprocesses (PrefectHQ/orion#2197)
When integration testing the `SubprocessFlowRunner` in various virtual
environment setups (`venv`, `virtualenv`), we were making a pretty major
assumption that the developer has `prefect` installed globally in their
`site-packages`.  This may not be the case, as some engineers (cough, myself)
set `PIP_REQUIRE_VIRTUALENV=1` and never install packages globally.

This means that I was never able to run the full test suite, as these tests
would always fail.  Also, it means that if you had an old verison of `prefect`
installed globally, you wouldn't be running the code you are currently working
on.

Introducing the conditional installation here is a compromise: if you do have
`prefect` installed globally and it is the right version (like it will be in
CI), then we take the fast route and just use it via `--system-site-packages`;
otherwise, we install it from your working tree.  We don't need the `[dev]`
extras (and in fact, we _should not_ use them), which brought the installation
time down to ~20s for me.  That's acceptable for running the full test suite,
especially since we can now run in parallel.